requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
Django==3.1.1
django-admin-numeric-filter==0.1.6
filelock==3.0.12
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
pytz==2020.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.31
wrapt==1.12.1
